+++ changelog.md
@@ -1,3 +1,6 @@
+## Forestry 2.1.1
+- Fix IIndividualHandlerItem methods returning individuals for empty stacks that were decremented
+
 ## Forestry 2.1.0
 - Fixed typo that cut off black order on the right side of the Portable Analyzer GUI
 - Add IActivityType bee chromosome for more complex sleep patterns (#97)